+++ CHANGES.md
@@ -137,6 +137,7 @@ Callable {
 * Fix the conflict between the logging kernel and the JDK threadpool plugin.
 * Fix the thread safety bug of finishing operation for the span named "SpringCloudGateway/sendRequest"
 * Fix NPE in guava-eventbus-plugin.
+* Add WebSphere Liberty 23.x plugin
 
 #### Documentation